+++ R/COSMIC.v3.1.R
@@ -6,13 +6,18 @@
 #'   saved in \code{data-raw/COSMIC.v3.1/data/}. \cr Populated by
 #'   \code{data-raw/COSMIC.v3.1/code/generate-COSMIC.v3.1-genome-sigs.R}.
 #'
-#' @format A list with a single element, \code{genome}, which is a
-#'   list containing:
+#' @format A list with one element, \code{signature}, which in turn is a list
+#' with elements:
+#'
+#' \describe{
+#'   \item{genome}{A list with the elements:
+#'
 #'   \describe{
 #'     \item{SBS96}{Strand-agnostic single-base substitutions in trinucleotide context.}
 #'     \item{SBS192}{Transcriptionally stranded single-base substitutions in trinucleotide context.}
 #'     \item{DBS78}{Strand-agnostic doublet-base substitutions.}
 #'     \item{ID}{Strand-agnostic indels.}
+#'   }}
 #'   }
 #'
 #' @section Remark: